refactor: extract normalizeOptions from executeRequest
